Fix issue with corrupted Jacoco coverage file (#58)
* Fix issue with corrupted Jacoco coverage file

Builds keep failing with:

```
* What went wrong:
Execution failed for task ':jacocoTestReport'.
> Unable to read execution data file /home/runner/work/creek-system-test-gradle-plugin/creek-system-test-gradle-plugin/build/jacoco/test.exec
```

Not exactly sure what is wrong with the test.exec file, but internet searches suggest it may be truncated due to JVMs exiting prematurely.

After lots of messing around the contained change works.  The `pkill` command doesn't look to shutdown Gradle Daemons often, but the splitting of the run into separate `check` and a `coveralls` executions seems to have done the trick.