Capture coverage metrics from system tests
fixes: https://github.com/creek-service/creek-system-test/issues/172

Adds a `systemTestPrepareCoverage` task, which prepares a mount directory containing the Jacoco agent jar.

Adds a `SystemTestCoverageExtension`, which extends the `SystemTest` task _if_ `Jacoco` plugin is applied. This will pass args to the test executor to add a read-only mount containing the Jacoco agent jar, and a writeable mount for the coverage results, then set `JAVA_TOOL_OPTIONS` to install the Jacoco agent in the service's JVM.