Avoid calling executor with empty debug args (#130)
If you supply a service name to debug, or just a service instance to debug, the plugin was passing the other args with an empty value.

For example, setting a service name to debug would result in the plugin passing `--debug-service-instance=`, i.e. an empty instance list.

Let's avoid that and just not set the argument if the list is empty.