Use new debug env (#131)
Enhance the plugin to make use of the new `--debug-env` option supported by the system test executor to pass in the `JAVA_TOOL_OPTIONS` required for supporting debugging of services.

This fixes a bug around debugging when there is more than one service-under-test in the test suite. Any service _not_ being debugged would fail to start as its `JAVA_TOOL_OPTIONS` incorrectly included the debug agent _and_ the agent line still had a placeholder for the debug port.

A workaround for this bug was to debug all services when you need to debug any.

BREAKING CHANGE: this is a braking change, not compatible with previous versions of the system test executor.  However, still in Alpha, so such changes allowed.